refactor: 테이블 수정
