+++ README.md
@@ -2,16 +2,11 @@
 
 ## 쿠폰 정책
 
-1. 금액 할인
-    1. ~이상 주문 시 적용 가능
-    2. 쿠폰에 정해진 액수만큼 할인
-    3. 사용자 별로 쿠폰 관리
-2. 퍼센트 할인
-    1. ~이상 주문 시 적용 가능
-    2. 정해진 퍼센트만큼 할인
-    3. 사용자 별로 쿠폰 관리
-3. 배송비 할인
-    1. 배송비 무료
+금액 할인
+
+1. ~이상 주문 시 적용 가능
+2. 쿠폰에 정해진 액수만큼 할인
+3. 사용자 별로 쿠폰 관리
 
 ## 백엔드 요구사항
 
@@ -195,10 +190,11 @@ HTTP/1.1 201 Created
       "quantity": 2
     }
   ],
-   "totalProductAmount" 55000,
+  "totalProductAmount"
+  55000,
   // 배달비를 포함하지 않은 상품 금액만 전부 더한 금액?
-   "deliveryAmount": 2000,
-   "discountedProductAmount": 3000,
+  "deliveryAmount": 2000,
+  "discountedProductAmount": 3000,
   // 할인 금액을 할 건지 or 할인된 총 금액을 할 건지
   "address": "서울특별시 송파구 ..."
 }
