+++ src/main/java/cart/domain/Coupon.java
@@ -1,12 +1,36 @@
 package cart.domain;
 
 import cart.domain.vo.Amount;
+import cart.exception.BusinessException;
 
-public interface Coupon {
+public class Coupon {
 
-    Amount calculateProduct(Amount productAmount);
+    private final Long id;
+    private final String name;
+    private final Amount discountAmount;
+    private final Amount minAmount;
+    private final boolean isUsed;
 
-    Amount calculateDelivery(Amount deliveryAmount);
+    public Coupon(final Long id, final String name, final Amount discountAmount, final Amount minAmount,
+        final boolean isUsed) {
+        this.id = id;
+        this.name = name;
+        this.discountAmount = discountAmount;
+        this.minAmount = minAmount;
+        this.isUsed = isUsed;
+    }
 
-    Coupon use();
+    public Amount calculateProduct(final Amount productAmount) {
+        if (isUsed) {
+            throw new BusinessException("이미 사용한 쿠폰입니다.");
+        }
+        if (productAmount.isLessThan(minAmount)) {
+            throw new BusinessException("총 금액은 최소금액보다 커야합니다.");
+        }
+        return productAmount.minus(discountAmount);
+    }
+
+    public Coupon use() {
+        return new Coupon(this.id, this.name, this.discountAmount, this.minAmount, this.isUsed);
+    }
 }
