+++ src/main/java/cart/dto/CartItemResponse.java
@@ -3,21 +3,25 @@ package cart.dto;
 import cart.domain.CartItem;
 
 public class CartItemResponse {
+
     private Long id;
     private int quantity;
     private ProductResponse product;
 
-    private CartItemResponse(Long id, int quantity, ProductResponse product) {
+    public CartItemResponse() {
+    }
+
+    private CartItemResponse(final Long id, final int quantity, final ProductResponse product) {
         this.id = id;
         this.quantity = quantity;
         this.product = product;
     }
 
-    public static CartItemResponse of(CartItem cartItem) {
+    public static CartItemResponse of(final CartItem cartItem) {
         return new CartItemResponse(
-                cartItem.getId(),
-                cartItem.getQuantity(),
-                ProductResponse.of(cartItem.getProduct())
+            cartItem.getId(),
+            cartItem.getQuantity(),
+            ProductResponse.of(cartItem.getProduct())
         );
     }
