+++ src/main/java/cart/dto/ProductResponse.java
@@ -4,10 +4,13 @@ import cart.domain.Product;
 
 public class ProductResponse {
 
-    private final Long id;
-    private final String name;
-    private final int price;
-    private final String imageUrl;
+    private Long id;
+    private String name;
+    private int price;
+    private String imageUrl;
+
+    public ProductResponse() {
+    }
 
     private ProductResponse(final Long id, final String name, final int price, final String imageUrl) {
         this.id = id;
