+++ src/test/java/cart/domain/CouponTest.java
@@ -8,13 +8,13 @@ import cart.exception.BusinessException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class AmountCouponTest {
+class CouponTest {
 
     @Test
     @DisplayName("총 상품 금액에 쿠폰을 적용한다.")
     void testCalculateProduct() {
         //given
-        final Coupon amountCoupon = new AmountCoupon(1L, "name", Amount.of(1_000), Amount.of(10_000), isUsed);
+        final Coupon amountCoupon = new Coupon(1L, "name", Amount.of(1_000), Amount.of(10_000), false);
         final Amount productAmount = Amount.of(15_000);
 
         //when
@@ -28,7 +28,7 @@ class AmountCouponTest {
     @DisplayName("총 상품 금액이 최소 금액보다 작으면 적용할 수 없다.")
     void testCalculateProductWhenProductAmountIsLessThanMinAmount() {
         //given
-        final Coupon amountCoupon = new AmountCoupon(1L, "name", Amount.of(1_000), Amount.of(10_000), isUsed);
+        final Coupon amountCoupon = new Coupon(1L, "name", Amount.of(1_000), Amount.of(10_000), false);
         final Amount productAmount = Amount.of(9_000);
 
         //when
